Normalize helper functions
